+++ packages/yourloops/layout/patient-layout.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023, Diabeloop
+ * Copyright (c) 2022-2025, Diabeloop
  *
  * All rights reserved.
  *
@@ -33,7 +33,7 @@ import { PatientCaregiversPage } from '../pages/patient/caregivers/patient-careg
 import { TeamContextProvider } from '../lib/team'
 import { DashboardLayout } from './dashboard-layout'
 import { InvalidRoute } from '../components/invalid-route'
-import { ProfilePage } from '../pages/profile/profile-page'
+import { UserAccountPage } from '../pages/user-account/user-account-page'
 import { NotificationsPage } from '../pages/notifications/notifications-page'
 import { AppUserRoute } from '../models/enums/routes.enum'
 import PatientUtils from '../lib/patient/patient.util'
@@ -47,7 +47,7 @@ export const PatientLayout: FC = () => {
       <DashboardLayout>
         <Routes>
           <Route path={AppUserRoute.NotFound} element={<InvalidRoute />} />
-          <Route path={AppUserRoute.Preferences} element={<ProfilePage />} />
+          <Route path={AppUserRoute.UserAccount} element={<UserAccountPage />} />
           <Route path={AppUserRoute.Notifications} element={<NotificationsPage />} />
           <Route path={AppUserRoute.Caregivers} element={<PatientCaregiversPage />} />
           <Route path={AppUserRoute.CareTeamSettings} element={<CareTeamSettingsPage />} />