+++ packages/yourloops/test/integration/pages/user-account/user-account.hcp.spec.tsx
@@ -29,12 +29,10 @@ import { renderPage } from '../../utils/render'
 import { loggedInUserEmail, loggedInUserId, mockAuth0Hook } from '../../mock/auth0.hook.mock'
 import { buildAvailableTeams, mockTeamAPI, myThirdTeamName } from '../../mock/team.api.mock'
 import { mockNotificationAPI } from '../../mock/notification.api.mock'
-import { fireEvent, screen, waitFor, within } from '@testing-library/react'
+import { waitFor } from '@testing-library/react'
 import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
 import { mockPatientApiForHcp } from '../../mock/patient.api.mock'
-import { checkHcpProfilePage, checkPasswordChangeRequest } from '../../assert/profile.assert'
-import userEvent from '@testing-library/user-event'
-import { type Profile } from '../../../../lib/auth/models/profile.model'
+import { type UserAccount } from '../../../../lib/auth/models/user-account.model'
 import { type Settings } from '../../../../lib/auth/models/settings.model'
 import { CountryCodes } from '../../../../lib/auth/models/country.model'
 import { LanguageCodes } from '../../../../lib/auth/models/enums/language-codes.enum'
@@ -51,9 +49,13 @@ import {
 } from '../../use-cases/app-main-layout-visualisation'
 import { ConfigService } from '../../../../lib/config/config.service'
 import ErrorApi from '../../../../lib/error/error.api'
+import { testHcpUserInfoUpdate, testPasswordChangeRequest } from '../../use-cases/user-account-management'
+import { AppUserRoute } from '../../../../models/enums/routes.enum'
 
-describe('Profile page for hcp', () => {
-  const profile: Profile = {
+describe('User account page for hcp', () => {
+  const userAccountRoute = AppUserRoute.UserAccount
+
+  const account: UserAccount = {
     email: 'djamal@alatete.com',
     firstName: 'Djamal',
     lastName: 'Alatete',
@@ -72,18 +74,18 @@ describe('Profile page for hcp', () => {
   beforeAll(() => {
     mockAuth0Hook()
     mockAuthApi()
-    mockUserApi().mockUserDataFetch({ profile, preferences, settings })
+    mockUserApi().mockUserDataFetch({ account, preferences, settings })
     mockNotificationAPI()
     mockDirectShareApi()
     mockTeamAPI()
     mockPatientApiForHcp()
   })
 
-  it('should render profile page for a French HCP and be able to edit his profile and change his password', async () => {
+  it('should render user account page for a French HCP and be able to edit his profile and change his password', async () => {
     jest.spyOn(ErrorApi, 'sendError').mockResolvedValue(null)
 
-    const expectedProfile = {
-      ...profile,
+    const expectedUserAccount = {
+      ...account,
       firstName: 'Jean',
       lastName: 'Talue',
       fullName: 'Jean Talue',
@@ -93,7 +95,7 @@ describe('Profile page for hcp', () => {
     const appMainLayoutParams: AppMainLayoutHcpParams = {
       footerHasLanguageSelector: false,
       headerInfo: {
-        loggedInUserFullName: `${profile.lastName} ${profile.firstName}`,
+        loggedInUserFullName: `${account.lastName} ${account.firstName}`,
         teamMenuInfo: {
           selectedTeamName: myThirdTeamName,
           isSelectedTeamPrivate: false,
@@ -103,74 +105,32 @@ describe('Profile page for hcp', () => {
     }
     const expectedPreferences = { displayLanguageCode: 'en' as LanguageCodes }
     const expectedSettings = { units: { bg: Unit.MilligramPerDeciliter }, country: CountryCodes.Austria }
-    const updateProfileMock = jest.spyOn(UserApi, 'updateProfile').mockResolvedValue(expectedProfile)
+    const updateUserAccountMock = jest.spyOn(UserApi, 'updateUserAccount').mockResolvedValue(expectedUserAccount)
     const updatePreferencesMock = jest.spyOn(UserApi, 'updatePreferences').mockResolvedValue(expectedPreferences)
     const updateSettingsMock = jest.spyOn(UserApi, 'updateSettings').mockResolvedValue(expectedSettings)
 
-    const router = renderPage('/preferences')
+    const router = renderPage(userAccountRoute)
     await waitFor(() => {
-      expect(router.state.location.pathname).toEqual('/preferences')
+      expect(router.state.location.pathname).toEqual(userAccountRoute)
     })
 
     await testAppMainLayoutForHcp(appMainLayoutParams)
-    const fields = checkHcpProfilePage()
-    const saveButton = screen.getByRole('button', { name: 'Save' })
 
-    expect(fields.firstNameInput).toHaveValue(profile.firstName)
-    expect(fields.lastNameInput).toHaveValue(profile.lastName)
-    expect(fields.unitsSelect).toHaveTextContent(settings.units.bg)
-    expect(fields.languageSelect).toHaveTextContent('Français')
-    expect(fields.hcpProfessionSelect).toHaveTextContent('Diabetologist')
-    expect(saveButton).toBeDisabled()
+    await testHcpUserInfoUpdate()
 
-    fireEvent.mouseDown(within(screen.getByTestId('profile-local-selector')).getByRole('combobox'))
-    fireEvent.click(screen.getByRole('option', { name: 'English' }))
-
-    fireEvent.mouseDown(within(screen.getByTestId('profile-units-selector')).getByRole('combobox'))
-    fireEvent.click(screen.getByRole('option', { name: Unit.MilligramPerDeciliter }))
-
-    fireEvent.mouseDown(within(screen.getByTestId('dropdown-profession-selector')).getByRole('combobox'))
-    fireEvent.click(screen.getByRole('option', { name: 'Nurse' }))
-
-    fireEvent.mouseDown(within(screen.getByTestId('country-selector')).getByRole('combobox'))
-    fireEvent.click(screen.getByRole('option', { name: 'Japan' }))
-
-    fireEvent.mouseDown(within(screen.getByTestId('country-selector')).getByRole('combobox'))
-    fireEvent.click(screen.getByRole('option', { name: 'United Kingdom' }))
-
-    fireEvent.mouseDown(within(screen.getByTestId('country-selector')).getByRole('combobox'))
-    fireEvent.click(screen.getByRole('option', { name: 'Austria' }))
-
-    await userEvent.clear(fields.firstNameInput)
-    await userEvent.clear(fields.lastNameInput)
-    await userEvent.type(fields.firstNameInput, 'Jean')
-    await userEvent.type(fields.lastNameInput, 'Talue')
-
-    expect(saveButton).not.toBeDisabled()
-    await userEvent.click(saveButton)
-
-    expect(saveButton).toBeDisabled()
-    expect(screen.getByRole('alert')).toBeVisible()
     expect(updatePreferencesMock).toHaveBeenCalledWith(loggedInUserId, expectedPreferences)
-    expect(updateProfileMock).toHaveBeenCalledWith(loggedInUserId, expectedProfile)
+    expect(updateUserAccountMock).toHaveBeenCalledWith(loggedInUserId, expectedUserAccount)
     expect(updateSettingsMock).toHaveBeenCalledWith(loggedInUserId, expectedSettings)
 
-    const profileUpdateSuccessfulSnackbar = screen.getByTestId('alert-snackbar')
-    expect(profileUpdateSuccessfulSnackbar).toHaveTextContent('Profile updated')
-
-    const profileUpdateSuccessfulSnackbarCloseButton = within(profileUpdateSuccessfulSnackbar).getByTitle('Close')
-
-    await userEvent.click(profileUpdateSuccessfulSnackbarCloseButton)
-
-    await checkPasswordChangeRequest(loggedInUserEmail)
+    await testPasswordChangeRequest(loggedInUserEmail)
   })
 
   it('should show the banner and change its content when the banner is enabled and language is changed', async () => {
     jest.spyOn(ConfigService, 'isBannerEnabled').mockReturnValue(true)
 
-    const router = renderPage('/preferences')
+    const router = renderPage(userAccountRoute)
     await waitFor(() => {
-      expect(router.state.location.pathname).toEqual('/preferences')
+      expect(router.state.location.pathname).toEqual(userAccountRoute)
     })
 
     await testBannerLanguageUpdate()