+++ CHANGELOG.md
@@ -28,6 +28,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3690 Update basal tooltip style in Daily view
 - YLP-3700 Update time change tooltip style in Daily view
 - YLP-3702 Update Time In Loop Mode stat style
+- YLP-3707 User account page new design
 - YLP-3710 Move the "EVENTS" section above the glycemia curve
 - YLP-3730 Delete "Avg. Daily" in dashboard and trends view to lighten all the frame
 - YLP-3731 Improve the calendar object in trend and daily view